refactor(repl-sdk): no barrel for the module fs either
Same rule as #2221: no index that re-exports, and reach each module directly.

    repl-sdk/fs           VFS, createSourceHook
    repl-sdk/fs/install   Installer
    repl-sdk/fs/npm       getTar
    repl-sdk/fs/url       npmUrl, specifierUrl, virtualUrl, typeFor

`src/index.js` already imported the modules directly, so only the browser
tests changed.
