--- packages/repl-sdk/src/fs/index.js
@@ -1,16 +1,0 @@
-export { PROJECT_PREFIX, releaseEntry, writeEntry } from './entry.js';
-export { Installer } from './install.js';
-export { clearTarCache, getTar } from './npm-tar.js';
-export { clearFs, installer, vfs } from './store.js';
-export {
-  NPM_PREFIX,
-  npmUrl,
-  parseNpmUrl,
-  parseVirtualUrl,
-  specifierUrl,
-  typeFor,
-  VIRTUAL_PREFIX,
-  virtualUrl,
-} from './url.js';
-export { createSourceHook, VFS } from './vfs.js';
-export { virtualModuleSource } from './virtual.js';
