+++ packages/repl-sdk/src/fs/npm-tar.d.ts
@@ -1,0 +1,4 @@
+import type { UntarredPackage } from '../types.ts';
+
+export function getTar(name: string, version: string): Promise<UntarredPackage>;
+export function clearTarCache(): void;
